Rename method: `add_user_to_chat` -> `add_users_to_chat`
