+++ tests/client/chats_api/test_remove_users.py
